kata: 7kyu-php PrinterErrors
