+++ tests/K7_PrinterErrors/DataProvider.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace tests\K7_PrinterErrors;
+
+class DataProvider
+{
+    public function data()
+    {
+        return [
+            ["aaaaaaaaaaaaaaaabbbbbbbbbbbbbbbbbbmmmmmmmmmmmmmmmmmmmxyz", "3/56"],
+            ["kkkwwwaaaaaaaaaaaaaabbbbbbbbbbbbbbbbbbmmmmmmmmmmmmmmmmmmmxyz", "6/60"],
+            ["kkkwwwaaaaaaaaaaaaaabbbbbbbbbbbbbbbbbbmmmmmmmmmmmmmmmmmmmxyzuuuuu", "11/65"],
+        ];
+    }
+}
